Add wyag ls-tree command.
ls-tree takes as input the SHA-1 of a tree object, and prints its
contents.

A tree is how git links blobs to file paths, and is formatted like this:
[mode] <space 0x20> [path] <null 0x00> [SHA-1]
[mode] <space 0x20> [path] <null 0x00> [SHA-1]
[mode] <space 0x20> [path] <null 0x00> [SHA-1]

where each line identifies either a blob (when "path" is a filename) or
another tree (when "path" is a directory). "mode" is the file system
permissions of the file.