Add wyag checkout command.
wyag checkout, unlike git checkout, requires two arguments: a commit (or
tree) to checkout, and a directory to use as a destination. This
simplifies away all of the nice safeguards git has in place and just
demonstrates the checkout in principle.

The commit object's tree is used as a starting place, and we recurse
into it, generating files from blobs and directories from subtrees,
until we have iterated through every item in the tree. Then, the state
of that directory will be exactly as if you had run "git checkout
<commit-hash>" (without the .git/ directory and ignored files of
course).